+++ src/signatures.rs
@@ -900,6 +900,9 @@ signatures! {
     value = b"\x01\x00\x02" offset = 8, b"LP" offset = 34
     value = b"\x02\x00\x02" offset = 8, b"LP" offset = 34
 
+    format = FigletFont
+    value = b"flf2a"
+
     format = GoogleDraco
     value = b"DRACO"
 
@@ -983,6 +986,9 @@ signatures! {
 
     format = Truetype
     value = b"\x00\x01\x00\x00\x00"
+
+    format = TruetypeCollection
+    value = b"ttcf\x00"
 
     format = VirtualMachineDisk
     value = b"COWD\x01"